chore(utils): reorganize file structure
