+++ src/app/player/pipes/format-volume.spec.ts
@@ -1,4 +1,4 @@
-import { testUtils } from 'src/utils/test';
+import { testUtils } from 'src/app/utils/test';
 import { FormatVolumePipe } from './format-volume';
