+++ src/app/tracklists/components/track-card.spec.ts
@@ -3,7 +3,7 @@ import { async, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { SharedModule } from 'src/app/shared';
-import { testUtils } from 'src/utils/test';
+import { testUtils } from 'src/app/utils/test';
 import { createTrack } from '../models/track';
 import { TrackCardComponent } from './track-card';
