+++ src/app/tracklists/state/selectors.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { Store, StoreModule } from '@ngrx/store';
 import { TRACKS_PER_PAGE } from 'src/constants';
-import { testUtils } from 'src/utils/test';
+import { testUtils } from 'src/app/utils/test';
 import { TracklistRecord } from '../models/tracklist';
 import { TracklistActions } from '../tracklist-actions';
 import { getCurrentTracklist, getTracklists, getTracks, getTracksForCurrentTracklist } from './selectors';
