+++ src/app/users/user-actions.spec.ts
@@ -1,4 +1,4 @@
-import { testUtils } from 'src/utils/test';
+import { testUtils } from 'src/app/utils/test';
 import { UserActions } from './user-actions';
 import { tracklistIdForUserLikes, tracklistIdForUserTracks } from './utils';
